(doc): Added a note on npm run dev (#565)
If port 8080 is in use on the machine npm run dev will fail without too much information to go on. reading through issues it looked like several people has experienced this and the first comment is always if port 8080 is in use. This edit might get more people going faster.